fixed comments bugs - updated on post_details.html, comments.js Relates #38
