+++ public/index.html
@@ -189,7 +189,7 @@
         <img class='side-image' src="./images/side-background.jpg" alt="">
 
         <div class="forms">
-          <div class="login" style="display: block;">
+          <div class="login" style="display: none;">
             <h2>Log In</h2>
             <form class="form-login">
               <div class="form-group">
