+++ server/routes/users.js
@@ -6,6 +6,6 @@ router.post('/login', userControllers.login);
 router.post('/signup', userControllers.signup);
 router.get('/auth-user', middlewares.authUser);
 router.get('/check-user', middlewares.checkUser);
-router.get('*', middlewares.checkUser);
+// router.get('*', middlewares.checkUser);
 
 module.exports = router;
